Expose function to convert method responses to models
Use that function in webhook updates.
