+++ test/ex_gram/cast_test.exs
@@ -1,0 +1,5 @@
+defmodule ExGram.CastTest do
+  use ExUnit.Case, async: true
+
+  doctest ExGram.Cast
+end
